feat: not count staff tickets to totals
