+++ src/components/eventos/modal/TicketsTable.tsx
@@ -27,6 +27,7 @@ const TicketsTable = () => {
       <TableBody>
         {tickets.map((ticket, index) => {
           const { icon, text } = iconsAndTexts[ticket.type];
+          if (ticket.type === 'STAFF') return null;
           return (
             <TableRow
               key={ticket.type}
